assessment/EDA/requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
autogen-agentchat
